+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.0] - 2020-05-09
+### Added
+- `purge` command now supports `--recursive` flag
+- Internal improvements and refactoring
+
 ## [0.14.0] - 2020-05-03
 ### Changed
 - Consolidating more logic into `engine.spark`